requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.1
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
autopep8==1.5.6
backcall==0.2.0
bleach==3.3.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
colour==0.1.5
cycler==0.10.0
decorator==5.0.5
defusedxml==0.7.1
distlib==0.3.1
distro==1.5.0
dj-database-url==0.5.0
Django==3.1.5
django-autoslug==1.9.8
django-cors-headers==3.7.0
django-crispy-forms==1.10.0
djangorestframework==3.12.4
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
firebase-admin==4.5.3
future==0.18.2
gcloud==0.18.3
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jws==0.1.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
msgpack==1.0.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.5
oauth2client==3.0.0
openpyxl==3.0.5
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pdfminer==20191125
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.10.0
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.15.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pyee==8.1.0
Pygments==2.8.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.1
pyppeteer==0.2.5
pyrsistent==0.17.3
python-bidi==0.4.2
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2020.5
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
reportlab==3.5.59
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
styleframe==3.0.6
tabula-py==2.2.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
xlrd==1.2.0
